backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.3
colorama==0.4.6
deprecated==1.2.18
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.1
h11==0.16.0
idna==3.15
limits==3.13.0
packaging==24.2
pydantic==2.13.4
pydantic_core==2.46.4
python-dotenv==1.2.2
slowapi==0.1.9
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.46.0
wrapt==1.17.0
